Rename FabButton to FAB
