+++ packages/experimental/Button/src/FAB/FAB.tsx
@@ -1,24 +1,24 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import { View } from 'react-native';
-import { fabButtonName, FabButtonType } from './FabButton.types';
+import { fabName, FABType } from './FAB.types';
 import { Text } from '@fluentui-react-native/experimental-text';
 import { compose, UseSlots } from '@fluentui-react-native/framework';
 import { Icon } from '@fluentui-react-native/icon';
 import { ButtonCoreProps } from '../Button.types';
 
-export const FabButton = compose<FabButtonType>({
-  displayName: fabButtonName,
+export const FAB = compose<FABType>({
+  displayName: fabName,
   slots: {
     root: View,
     icon: Icon,
     content: Text,
   },
-  render: (_userProps: ButtonCoreProps, _useSlots: UseSlots<FabButtonType>) => {
+  render: (_userProps: ButtonCoreProps, _useSlots: UseSlots<FABType>) => {
     return (_final: ButtonCoreProps, ..._children: React.ReactNode[]) => {
       return null;
     };
   },
 });
 
-export default FabButton;
+export default FAB;
