+++ packages/experimental/Button/src/FAB/FABTokens.ts
@@ -3,7 +3,7 @@ import { globalTokens } from '@fluentui-react-native/theme-tokens';
 import { TokenSettings } from '@fluentui-react-native/use-styling';
 import { ButtonCoreTokens } from '../Button.types';
 
-export const defaultFabButtonTokens: TokenSettings<ButtonCoreTokens, Theme> = () =>
+export const defaultFABTokens: TokenSettings<ButtonCoreTokens, Theme> = () =>
   ({
     borderRadius: globalTokens.corner.radius.circle,
     minHeight: 56,
