+++ packages/experimental/Button/src/FAB/index.ts
@@ -1,0 +1,2 @@
+export * from './FAB.types';
+export * from './FAB';
