--- packages/experimental/Button/src/FabButton/index.ts
@@ -1,2 +1,0 @@
-export * from './FabButton.types';
-export * from './FabButton';
