+++ packages/experimental/Button/src/index.ts
@@ -1,5 +1,5 @@
 export * from './Button.types';
 export * from './Button';
 export * from './CompoundButton';
-export * from './FabButton';
+export * from './FAB';
 export * from './ToggleButton';
